undo AF2 changes, move modified functions to structure_prediction
